requirements.txt
tensorflow
matplotlib
numpy
scikit-learn
pillow